Socket IO client enhancement with connect() method
Updated the Socket client service, with a connect() method. Moved state
redirect out of from Socket service.  Added the Authentication.user
check to the Chat client controller, and added a check to make sure the
Socket client service is connected to the server; if not, then connect
using the new connect() method.

Had to do a hard reset from 0.4.0 due to conflicts when merging and
pushing to remote.